+++ tests/testreduce/rtTestWrapper.js
@@ -1,7 +1,7 @@
 'use strict';
 
 /* This is a wrapper around the roundtrip testing script */
-var apiServer = require('../apiServer.js');
+var serviceWrapper = require('../serviceWrapper.js');
 var rtTest = require('../../bin/roundtrip-test.js');
 
 // If we've started a Parsoid server, cache the URL so that subsequent calls
@@ -22,20 +22,13 @@ var runRoundTripTest = function(config, test) {
 		return _run(test);
 	} else {
 		// If no Parsoid server was passed, start our own
-		var p = apiServer.startParsoidServer({
-			serverArgv: [
-				// We want the cluster master so that timeouts on stuck titles
-				// lead to a restart.
-				'--num-workers', '1',
-				'--config', config.parsoidConfig,
-			],
-			quiet: true,
+		return serviceWrapper.runServices({
+			skipMock: true,
+			localsettings: config.parsoidConfig,
 		}).then(function(ret) {
-			parsoidURL = ret.url;
+			parsoidURL = ret.parsoidURL;
 			return _run(test);
 		});
-		apiServer.exitOnProcessTerm();
-		return p;
 	}
 };